whitespace/app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.4.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.8.2
    # via streamlit
cachetools==5.4.0
    # via streamlit
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
distro==1.9.0
    # via openai
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jiter==0.5.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
langchain-core==0.2.33
    # via langchain-openai
langchain-openai==0.1.22
    # via -r requirements.in
langsmith==0.1.99
    # via langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
narwhals==1.4.2
    # via altair
numpy==2.0.1
    # via
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
openai==1.41.0
    # via langchain-openai
orjson==3.10.7
    # via langsmith
packaging==24.1
    # via
    #   altair
    #   langchain-core
    #   streamlit
pandas==2.2.2
    # via streamlit
pillow==10.4.0
    # via streamlit
protobuf==5.27.3
    # via streamlit
pyarrow==17.0.0
    # via streamlit
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.20.1
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via langchain-core
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.7.24
    # via tiktoken
requests==2.32.3
    # via
    #   langsmith
    #   streamlit
    #   tiktoken
rich==13.7.1
    # via streamlit
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
streamlit==1.37.0
    # via -r requirements.in
tenacity==8.5.0
    # via
    #   langchain-core
    #   streamlit
tiktoken==0.7.0
    # via langchain-openai
toml==0.10.2
    # via streamlit
tornado==6.4.1
    # via streamlit
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   altair
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   streamlit
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
